Externalizing DataTypeSelector in order to be re-used elsewhere
