+++ kie-wb-common-react/boxed-expression-editor/src/components/EditExpressionMenu/index.ts
@@ -14,4 +14,5 @@
  * limitations under the License.
  */
 
+export * from "./DataTypeSelector";
 export * from "./EditExpressionMenu";
